+++ httpdocs/js/user.js
@@ -6,414 +6,7 @@ export default class User extends Router {
   constructor(title, footer, routes) {
     super(title, footer, routes);
     this.routes.push({url: '/settings', setup: settingsPage});
-    this.routes.push({url: '/my-projects', setup: myProjectsPage});
-    this.page = 1;
-    this.search = '';
-    this.sort = 'default';
-    this.delay = false;
     let that = this;
-    function myProjectsPage() {
-      that.page = new URL(document.location.href).searchParams.get('p') ?
-        parseInt(new URL(document.location.href).searchParams.get('p')) : 1;
-      that.search = new URL(document.location.href).searchParams.get('search') ?
-        (new URL(document.location.href).searchParams.get('search')).toString() : that.search;
-      that.sort = new URL(document.location.href).searchParams.get('sort') ?
-        (new URL(document.location.href).searchParams.get('sort')).toString() : that.sort;
-
-      // we need to be logged in to view this page
-      if (!that.password || !that.email)
-       return false;
-      const template = document.createElement('template');
-      const projectsTable =
-        `<section class="section" data-content="my-projects" style="padding: 0">
-          <div class="table-container my-projects-table mx-auto">
-            <div class="search-bar" style="max-width: 280px; padding-bottom: 20px;">
-              <div class="control has-icons-right">
-                <input class="input is-small" id="my-projects-search-input" type="text" placeholder="Search for projects...">
-                <span class="icon is-small is-right is-clickable" id="my-projects-search-click">
-                  <i class="fas fa-search" id="my-projects-search-icon"></i>
-                </span>
-              </div>
-            </div>
-            <table class="table is-striped is-hoverable">
-              <thead>
-                <tr>
-                  <th class="is-clickable column-title" id="my-projects-sort-viewed" title="Popularity"
-                    style="text-align:center; min-width: 65px;">
-                    <i class="fas fa-chart-column"></i>
-                    <i class="sort-icon fa-solid fa-sort-down" style="display: none;"></i>
-                  </th>
-                  <th class="is-clickable column-title" id="my-projects-sort-title" title="Title of the animation"
-                    style="min-width: 120px;">
-                    Title<i class="sort-icon fa-solid fa-sort-down" style="display: none;"></i>
-                  </th>
-                  <th class="is-clickable column-title" id="my-projects-sort-version" title="Webots release of the animation"
-                    style="min-width: 85px;">
-                    Version<i class="sort-icon fa-solid fa-sort-down" style="display: none;"></i>
-                  </th>
-                  <th class="is-clickable column-title" id="my-projects-sort-duration" title="Duration of the animation"
-                    style="text-align: right; min-width: 75px;">
-                    Duration<i class="sort-icon fa-solid fa-sort-down" style="display: none;"></i>
-                  </th>
-                  <th class="is-clickable column-title" id="my-projects-sort-size" title="Total size of the animation files"
-                    style="text-align: right; min-width: 75px;">
-                    Size<i class="sort-icon fa-solid fa-sort-down" style="display: none;"></i>
-                  </th>
-                  <th class="is-clickable column-title" id="my-projects-sort-uploaded" title="Upload date and time"
-                    style="text-align: right; min-width: 115px;">
-                    Uploaded<i class="sort-icon fa-solid fa-sort-down" style="display: none;"></i>
-                  </th>
-                </tr>
-              </thead>
-              <tbody>
-              </tbody>
-            </table>
-            <div class="empty-search" id="my-projects-empty-search" style="display: none;">
-              <i class="fas fa-xl fa-search" id="no-project-icon"
-                style="color: lightgrey; padding-right: 10px; position: relative; top: 12px;"></i>
-              <p id="my-projects-empty-search-text"></p>
-            </div>
-          </div>
-          <nav class="pagination is-small is-rounded mx-auto" role="navigation" aria-label="pagination"></nav>
-        </section>`;
-      template.innerHTML =
-      `<section class="section">
-        <div class="tile is-ancestor">
-          <div class="tile is-60 is-parent">
-            <div class="tile is-child box" style="overflow-X: auto">
-              <p class="title">My Projects</p>
-              <div class="content">
-                ${projectsTable}
-              </div>
-            </div>
-          </div>
-          <div class="tile is-vertical is-parent">
-            <div class="tile is-child box" id="my-projects-top">
-              <p class="title">Top Project <span id="my-projects-title" class="is-size-4 has-text-weight-normal"></span></p>
-              <div id="my-projects-top-container"></div>
-            </div>
-            <div class="tile is-child box">
-              <p class="title">Information</p>
-              <section class="section" id="my-projects-information" style="padding: 0;">
-              </section>
-            </div>
-          </div>
-        </div>
-      </section>`;
-      that.setup('settings', [], template.content);
-      listMyProjects();
-      initMyProjectsSearch();
-      initMyProjectsSort();
-      getUserStats();
-    }
-    function getUserStats() {
-      fetch('/ajax/animation/information.php', {method: 'post', body: JSON.stringify({user: that.id})})
-      .then(function(response) {
-        return response.json();
-      })
-      .then(function(data) {
-        if (data.error)
-          ModalDialog.run(`User information request error`, data.error);
-        else if (data.status !== 'no uploads') {
-          showTopProject(data);
-          showInformation(data);
-        }
-      });
-    }
-    function showInformation(data) {
-      document.getElementById('my-projects-information').innerHTML =
-        `<p style="padding-bottom: 10px;"><strong>First Upload: &nbsp; </strong>${data.firstUpload}</p>
-        <p style="padding-bottom: 10px;"><strong>Total Scenes: &nbsp; </strong>${data.totalScenes}</p>
-        <p style="padding-bottom: 10px;"><strong>Total Animations: &nbsp; </strong>${data.totalAnimations}</p>
-        <p><strong>Total Views: &nbsp; </strong>${data.totalViews}</p>`
-    }
-    function showTopProject(data) {
-      if (!that.topProjectWebotsView) {
-        that.topProjectWebotsView = document.createElement('webots-view');
-        that.topProjectWebotsView.id = 'my-projects-top-webots-view';
-        document.getElementById('my-projects-top-container').append(that.topProjectWebotsView);
-      }
-
-      document.getElementById('my-projects-title').innerHTML = data.topTitle;
-
-      const reference = window.location.origin + '/storage/' + data.topId;
-      if (data.topId[0] === 'S')
-        that.topProjectWebotsView.loadScene(`${reference}/scene.x3d`, that.isMobileDevice(), `${reference}/thumbnail.jpg`);
-      else
-        that.topProjectWebotsView.loadAnimation(`${reference}/scene.x3d`, `${reference}/animation.json`,
-          that.isMobileDevice(), false, `${reference}/thumbnail.jpg`);
-    }
-    function updateMyProjectsPagination(max) {
-      const hrefSort = that.sort && that.sort !== 'default' ? '?sort=' + that.sort : '';
-      const hrefSearch = that.search && that.search !== '' ? '?search=' + that.search : '';
-      let nav = document.querySelector(`section[data-content="my-projects"] > nav`);
-      let content = {};
-      const previousDisabled = (that.page === 1) ? ' disabled' : ` href="${(that.page === 2)
-        ? ('/my-projects') : ('/my-projects?p=' + (that.page - 1))}${hrefSort}${hrefSearch}"`;
-      const nextDisabled = (that.page === max) ? ' disabled' : ` href="my-projects?p=${that.page + 1}${hrefSort}${hrefSearch}"`;
-      const oneIsCurrent = (that.page === 1) ? ' is-current" aria-label="Page 1" aria-current="page"'
-        : `" aria-label="Goto page 1" href="my-projects${hrefSort}${hrefSearch}"`;
-      content.innerHTML =
-        `<a class="pagination-previous"${previousDisabled}>Previous</a>
-        <ul class="pagination-list"><li>
-        <a class="pagination-link${oneIsCurrent}>1</a></li>`;
-      for (let i = 2; i <= max; i++) {
-        if (i === that.page - 2 || (i === that.page + 2 && i !== max)) {
-          content.innerHTML += `<li><span class="pagination-ellipsis">&hellip;</span></li>`;
-          continue;
-        }
-        if (i < that.page - 2 || (i > that.page + 2 && i !== max))
-          continue;
-        if (i === that.page)
-          content.innerHTML += `<li><a class="pagination-link is-current" aria-label="Page ${i}"` +
-            ` aria-current="page">${i}</a></li>`;
-        else
-          content.innerHTML += `<li><a class="pagination-link" aria-label="Goto page ${i}"
-            href="/my-projects?p=${i}${hrefSort}${hrefSearch}">${i}</a></li>`;
-      }
-      content.innerHTML += `</ul>` + `<a class="pagination-next"${nextDisabled}>Next page</a>`;
-      nav.innerHTML = content.innerHTML;
-    }
-    function myProjectsRow(data) {
-      let size = data.size;
-      let unit;
-      if (size < 1024)
-        unit = 'bytes';
-      else if (size < 1024 * 1014) {
-        size = size / 1024;
-        unit = 'K';
-      } else if (size < 1024 * 1024 * 1024) {
-        size = size / (1024 * 1024);
-        unit = 'M';
-      } else {
-        size = size / (1024 * 1024 * 1024);
-        unit = 'G';
-      }
-      if (size < 100)
-        size = Math.round(10 * size) / 10;
-      else
-        size = Math.round(size);
-      size += ' <small>' + unit + '</small>';
-      let duration;
-      if (data.duration) {
-        let millisecond = data.duration % 1000;
-        let second = Math.trunc(data.duration / 1000) % 60;
-        let minute = Math.trunc(data.duration / 60000) % 60;
-        let hour = Math.trunc(data.duration / 3600000);
-        if (millisecond < 10)
-          millisecond = '00' + millisecond;
-        else if (millisecond < 100)
-          millisecond = '0' + millisecond;
-        duration = second + ':' + millisecond;
-        if (data.duration >= 60000) {
-          if (second < 10)
-            duration = '0' + duration;
-          duration = minute + ':' + duration;
-          if (data.duration > 3600000) {
-            if (minute < 10)
-              duration = '0' + duration;
-            duration = hour + duration;
-          }
-        }
-      } else
-        duration = '-';
-      const type = (data.duration === 0) ? 'scene' : 'animation';
-      const typeLetter = (data.duration === 0) ? 'S' : 'A';
-      const url = data.url.slice(0, data.url.lastIndexOf('/') + 1) + typeLetter + data.url.slice(data.url.lastIndexOf('/') + 1);
-      const thumbnailUrl = url.slice(0, url.lastIndexOf('/')) + '/storage' + url.slice(url.lastIndexOf('/')) + '/thumbnail.jpg';
-      const defaultThumbnailUrl = document.location.origin + '/images/thumbnail_not_available.jpg';
-      const versionUrl = `https://github.com/cyberbotics/webots/releases/tag/${data.version}`;
-      const style = ' style="color:#007acc"';
-      const tooltip = `Delete your ${type}`;
-      const deleteIcon = `<i${style} class="is-clickable far fa-trash-alt" id="my-projects-${data.id}" title="${tooltip}"></i>`;
-      const uploaded = data.uploaded.replace(' ', `<br>${deleteIcon} `);
-      const title = data.title === '' ? '<i>anonymous</i>' : data.title;
-      let row = `<td class="has-text-centered">${data.viewed}</td>`;
-      row += `<td>
-                <a class="table-title has-text-dark" href="${url}">${title}</a>
-                <div class="thumbnail">
-                  <div class="thumbnail-container">
-                    <img class="thumbnail-image" src="${thumbnailUrl}" onerror="this.src='${defaultThumbnailUrl}';"/>
-                    <p class="thumbnail-description">
-                      ${data.description}
-                      <div class="thumbnail-description-fade"/>
-                    </p>
-                  </div>
-                </div>
-              </td>`;
-      row += `<td><a class="has-text-dark" href="${versionUrl}" target="_blank"
-        title="View Webots release">${data.version}</a></td>`;
-      row += `<td class="has-text-right">${duration}</td>`;
-      row += `<td class="has-text-right">${size}</td><td class="has-text-right is-size-7">${uploaded}</td>`;
-      return row;
-    }
-    function listMyProjects() {
-      const pageLimit = 10;
-      const offset = (that.page - 1) * pageLimit;
-      fetch('/ajax/animation/list.php', {method: 'post',
-        body: JSON.stringify({offset: offset, limit: pageLimit, sortBy: that.sort, search: that.search, user: that.id})})
-        .then(function(response) {
-          return response.json();
-        })
-        .then(function(data) {
-          if (data.error)
-            ModalDialog.run(`User project listing error`, data.error);
-          else {
-            if (data.total === 0) {
-              const message = that.search || that.search !== '' ?
-                'Your search - <strong>' + that.search + '</strong> - did not match any projects.' :
-                'You have not uploaded any projects yet.';
-              const iconClassList = that.search || that.search !== '' ?
-                'fas fa-xl fa-search' : 'fas fa-xl fa-face-frown';
-              document.getElementById('no-project-icon').classList = iconClassList;
-              document.getElementById('my-projects-empty-search-text').innerHTML = message;
-              document.getElementById('my-projects-empty-search').style.display = 'flex';
-            } else
-              document.getElementById('my-projects-empty-search').style.display = 'none';
-            let line = ``;
-            for (let i = 0; i < data.animations.length; i++)
-              line += '<tr>' + myProjectsRow(data.animations[i]) + '</tr>';
-            let parent = that.content.querySelector(`section[data-content="my-projects"] > div > table > tbody`);
-            parent.innerHTML = line;
-            for (let i = 0; i < data.animations.length; i++) {
-              let type = data.animations[i].duration === 0 ? 'S' : 'A';
-              let node = parent.querySelector(`#my-projects-${data.animations[i].id}`);
-              if (node) {
-                let p = (data.animations.length === 1) ? that.page - 1 : that.page;
-                if (p === 0)
-                  p = 1;
-                node.addEventListener('click', function(event) { deleteMyProject(event, type, p); });
-              }
-            }
-            const total = (data.total === 0) ? 1 : Math.ceil(data.total / pageLimit);
-            updateMyProjectsPagination(total);
-          }
-        });
-    }
-    function initMyProjectsSort() {
-      if (that.sort && that.sort !== 'default') {
-        const columnTitle = document.getElementById('my-projects-sort-' + that.sort.split('-')[0]);
-        const sortIcon = columnTitle.querySelector('.sort-icon');
-        columnTitle.querySelector('.sort-icon').style.display = 'inline';
-        if (that.sort.split('-')[1] === 'asc' && sortIcon.classList.contains('fa-sort-down')) {
-          sortIcon.classList.toggle('fa-sort-down');
-          sortIcon.classList.toggle('fa-sort-up');
-        }
-      }
-      document.querySelectorAll('.column-title').forEach((title) => {
-        title.addEventListener('click', function(e) {
-          const sortIcon = title.querySelector('.sort-icon');
-          const previousSort = that.sort.split('-')[0];
-          let sort = title.id.split('-')[3];
-
-          if (previousSort === sort) {
-            sortIcon.classList.toggle('fa-sort-down');
-            sortIcon.classList.toggle('fa-sort-up');
-            sort += sortIcon.classList.contains('fa-sort-down') ? '-desc' : '-asc';
-          } else if (previousSort !== 'default') {
-            document.getElementById('my-projects-sort-' + previousSort).querySelector('.sort-icon').style.display = 'none';
-            if (sortIcon.classList.contains('fa-sort-up')) {
-              sortIcon.classList.toggle('fa-sort-down');
-              sortIcon.classList.toggle('fa-sort-up');
-            }
-            sort += '-desc';
-          } else
-            sort += '-desc';
-
-          title.querySelector('.sort-icon').style.display = 'inline';
-          that.sort = sort;
-          searchAndSortMyProjectsTable();
-        });
-      });
-    }
-    function initMyProjectsSearch() {
-      document.getElementById('my-projects-search-input').value = that.search;
-      document.getElementById('my-projects-search-input').addEventListener('keyup', function(event) {
-        if (!that.delay) {
-          that.delay = true;
-          setTimeout(() => {
-            that.search = document.getElementById('my-projects-search-input').value;
-            that.page = 1;
-            updateMyProjectsSearchIcon();
-            searchAndSortMyProjectsTable();
-            that.delay = false;
-          }, '300');
-        }
-      });
-      document.getElementById('my-projects-search-click').addEventListener('click', function(event) {
-        if (document.getElementById('my-projects-search-icon').classList.contains('fa-xmark')) {
-          document.getElementById('my-projects-search-input').value = '';
-          that.search = document.getElementById('my-projects-search-input').value;
-          that.page = 1;
-          updateMyProjectsSearchIcon();
-          searchAndSortMyProjectsTable();
-        }
-      });
-    }
-    function searchAndSortMyProjectsTable(isSearch) {
-      let url = new URL(document.location.origin + document.location.pathname);
-      if (that.page !== 1 && !isSearch)
-        url.searchParams.append('p', that.page);
-      else
-        that.page = 1;
-      if (that.sort && that.sort !== 'default')
-        url.searchParams.append('sort', that.sort);
-      if (that.search && that.search !== '')
-        url.searchParams.append('search', that.search);
-      window.history.replaceState(null, '', (url.pathname + url.search).toString());
-
-      listMyProjects();
-    }
-    function updateMyProjectsSearchIcon() {
-      console.log(that.search.length);
-      const searchIcon = document.getElementById('my-projects-search-icon');
-      if (searchIcon.classList.contains('fa-search') && that.search.length > 0) {
-        searchIcon.classList.remove('fa-search');
-        searchIcon.classList.add('fa-xmark');
-      } else if (searchIcon.classList.contains('fa-xmark') && that.search.length === 0) {
-        searchIcon.classList.add('fa-search');
-        searchIcon.classList.remove('fa-xmark');
-      }
-    }
-    function deleteMyProject(event, type, page) {
-      const animation = parseInt(event.target.id.substring(12));
-      const typeName = (type === 'A') ? 'animation' : 'scene';
-      const capitalizedTypeName = typeName.charAt(0).toUpperCase() + typeName.slice(1);
-      let dialog = ModalDialog.run(`Really delete ${typeName}?`, '<p>There is no way to recover deleted data.</p>', 'Cancel',
-        `Delete ${capitalizedTypeName}`, 'is-danger');
-      dialog.querySelector('form').addEventListener('submit', function(event) {
-        event.preventDefault();
-        dialog.querySelector('button[type="submit"]').classList.add('is-loading');
-        const content = {
-          method: 'post',
-          body: JSON.stringify({
-            type: type,
-            animation: animation,
-            user: that.id,
-            password: that.password
-          })
-        };
-        fetch('ajax/animation/delete.php', content)
-          .then(function(response) {
-            return response.json();
-          })
-          .then(function(data) {
-            dialog.close();
-            if (data.error)
-              ModalDialog.run(`${capitalizedTypeName} deletion error`, data.error);
-            else if (data.status === 1) {
-              let uploads = JSON.parse(window.localStorage.getItem('uploads'));
-              if (uploads !== null && uploads.includes(animation)) {
-                uploads.splice(uploads.indexOf(animation), 1);
-                if (uploads.length === 0)
-                  uploads = null;
-              }
-              window.localStorage.setItem('uploads', JSON.stringify(uploads));
-              that.load(`/my-projects${(page > 1) ? ('?p=' + page) : ''}`);
-            }
-          });
-      });
-    }
 
     function settingsPage() {
       // we need to be logged in to view this page
